Added files for simulating student pilots training at an airport.
Aircraft objects are individual planes which the students are flying, and they may take off, perform 'touch-and-gos', or land, each of which takes a certain amount of time on the runway, and the runway must be occupied by only one plane at a time. To accomplish that, the runway is simulated by a lock. The Airport method simply supplies the main method, which takes the number of planes in the simulation as the first argument and the length of a time of a 'time unit' in milliseconds.