Fixed/changed the lock classes.
Fixed the LockC class. Previously, the 'unlocked' value for the common variable was 0, but that is a value that should be available to threads as an id number, so the unlocked value was changed to -1. In LockB, the use of the AtomicInteger class was removed. Instead, volatile arrays are used and simply re-assigned to themselves whenever an element's value is updated.